Update parseAllowance, parseExpense, parseBudget methods
Added a check to the parseAllowance(), parseExpense() and parseBudget() methods in Parser.java to ensure that the following conditions are met for the user entered amount. The allowance, expense and budget amount is positive, less than 100000, and has at most two decimal places. This ensures that arbitrary large or small float amounts are rejected and not used.